requirements.txt
fastapi
uvicorn
python-gitlab
jinja2
python-multipart
requests